`bash-prg-hash`: remove block_api, inline code in lib.rs, add documentation
